# openapi.json
{
  "openapi": "3.0.1",
  "components": {},
  "info": {
    "title": "test_task",
    "description": "API for dealing with transactions history based on limits table",
    "contact": {
      "name": "Grigory Kovshov",
      "email": "barkev2009@bk.ru"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://{host}:{port}/api",
      "description": "Main server",
      "variables": {
        "host": {
          "default": "127.0.0.1",
          "description": "local host"
        },
        "port": {
          "default": "8000"
        }
      }
    }
  ],
  "paths": {
    "/limits": {
      "get": {
        "description": "Returns table of all limits",
        "responses": {
          "200": {
            "description": "A JSON of all limits",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "status": "success",
                    "data": [
                    {
                        "ID": 1,
                        "COUNTRY": "RUS",
                        "CUR": "RUB",
                        "MAX_LIMIT": 5000.0
                    },
                    {
                        "ID": 2,
                        "COUNTRY": "AUS",
                        "CUR": "USD",
                        "MAX_LIMIT": 5000.0
                    },
                      {
                        "ID": 3,
                        "COUNTRY": "ABH",
                        "CUR": "EUR",
                        "MAX_LIMIT": 5000.0
                      }
                    ]
                  }
                }
              }
            }
          },
          "500": {
            "description": "A failure to get all limits' table due to unknown error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "status": "failure",
                    "message": "ERROR_MESSAGE"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/limits/{id}": {
      "get": {
        "description": "Returns a limit record for a certain ID",
        "responses": {
          "200": {
            "description": "A limit record for a certain ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "status": "success",
                    "data": {
                      "ID": 1,
                      "COUNTRY": "RUS",
                      "CUR": "RUB",
                      "MAX_LIMIT": 5000.0
                    }
                  }
                }
              }
            }
          },
          "418": {
            "description": "A failure to get limit record possibly due to non-existent ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "status": "failure",
                    "message": "ID=444 does not exist"
                  }
                }
              }
            }
          },
          "500": {
            "description": "A failure to get limit record due to unknown error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "status": "failure",
                    "message": "ERROR_MESSAGE"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "Delete a limit record from the table",
        "responses": {
          "200": {
            "description": "Successfully deleted limit record",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "status": "success",
                    "message": "limit of id=4 deleted"
                  }
                }
              }
            }
          },
          "418": {
            "description": "Failure to delete limit record possibly due to non-existent ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "status": "failure",
                    "message": "limit of id=444 does not exist"
                  }
                }
              }
            }
          },
          "500": {
            "description": "A failure to delete limit record due to unknown error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "status": "failure",
                    "message": "ERROR_MESSAGE"
                  }
                }
              }
            }
          }
        }
      },
      "parameters": [
        {
          "name": "id",
          "description": "ID to get limit record of",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer"
          },
          "example": 1
        }
      ]
    },
    "/limits?id={id}&cur={currency}&country={country}&max_limit={max_limit}": {
      "post": {
        "description": "Posting a new limit record",
        "responses": {
          "200": {
            "description": "A new limit record",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "status": "success",
                    "message": "limits updated",
                    "result": {
                      "ID": 4,
                      "COUNTRY": "RUS",
                      "CUR": "RUB",
                      "MAX_LIMIT": 5000.0
                    }
                  }
                }
              }
            }
          },
          "418": {
            "description": "Failure to post new record due to incorrect inputs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "status": "failure",
                    "message": "limits not updated",
                    "result": {
                      "failure": "Failed to update limits. Please, check your input"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failure to post new record due to unknown error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "status": "failure",
                    "message": "ERROR_MESSAGE"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "Updating an existent limit record in the table",
        "responses": {
          "200": {
            "description": "Limit record successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "status": "success",
                    "message": "limits updated",
                    "result": {
                      "ID": 4,
                      "COUNTRY": "RUS",
                      "CUR": "RUB",
                      "MAX_LIMIT": 5000.0
                    }
                  }
                }
              }
            }
          },
          "418": {
            "description": "Failure to update record due to incorrect inputs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "status": "failure",
                    "message": "limits not updated",
                    "result": "check inputs"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failure to update record due to unknown error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "status": "failure",
                    "message": "ERROR_MESSAGE"
                  }
                }
              }
            }
          }
        }
      },
      "parameters": [
        {
          "name": "id",
          "description": "ID of the limit record (new or subject to update)",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer"
          },
          "example": 4
        },
        {
          "name": "currency",
          "description": "Currency of the limit record (new or subject to update)",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "RUB", "USD", "EUR"
            ]
          }
        },
        {
          "name": "country",
          "description": "Country of the limit record (new or subject to update)",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "RUS", "AUS", "ABH"
            ]
          }
        },
        {
          "name": "max_limit",
          "description": "Maximum limit of the limit record in corresponding currency (new or subject to update)",
          "in": "path",
          "required": true,
          "schema": {
            "type": "number",
            "example": 3000
          }
        }
      ]
    },
    "/history?id={id}&cur={currency}&country={country}&amount={amount}&date={date}": {
      "post": {
        "description": "Posting a new history record",
        "responses": {
          "200": {
            "description": "A new history record",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "status": "success",
                    "message": "history updated",
                    "result": {
                      "message": "history successfully updated"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failure to post new history record due to incorrect amount",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                      "status": "failure",
                      "message": "history not updated",
                      "result": {
                          "failure": "Sum within a current month (7001420.0 rub) exceeds max_limit (3000.0 rub)"
                      }
                  }
                }
              }
            }
          },
          "418": {
            "description": "Failure to post new history record due to incorrect input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                      "status": "failure",
                      "message": "history not updated",
                      "result": {
                          "failure": "Failed to update history. Please, check your inputs"
                      }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failure to post new history record due to unknown error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "status": "failure",
                    "message": "ERROR_MESSAGE"
                  }
                }
              }
            }
          }
        }
      },
      "parameters": [
        {
          "name": "id",
          "description": "ID of the new history record",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer"
          },
          "example": 4
        },
        {
          "name": "currency",
          "description": "Currency of the new history record",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "RUB", "USD", "EUR"
            ]
          }
        },
        {
          "name": "country",
          "description": "Country of the new history record",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "RUS", "AUS", "ABH"
            ]
          }
        },
        {
          "name": "amount",
          "description": "Amount of the new history record in corresponding currency",
          "in": "path",
          "required": true,
          "schema": {
            "type": "number",
            "example": 3000
          }
        },
        {
          "name": "date",
          "description": "Date of the new history record in format %Y-%m-%dT%H:%M:%S (or use 'now' for posting current time)",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "example": "2021-12-12T12:12:12"
          }
        }
      ]
    }
  }
}